Create README.md
An experienced and committed full-stack developer is required for the development of a mobile app. SUBTAP makes subcontracting in the construction industry simpler, fairer and more financially rewarding for everyone. A unique opportunity to be part of a company that will revolutionise the way thousands of people experience work. SUBTAP gives subcontractors higher earnings and more power. Employers pay less for an improved service and HM Treasury receives increased tax revenue. Further, SUBTAP helps improve site conditions and helps raise the standard of workmanship across the industry.

As Developer your role will be to oversee the development of mobile applications – initially for web, eventually native – from concept to delivery. You will be motivated and prepared to apply your skills and expertise in a variety of ways. As well as having a keen eye for good design, you will be self-reliant and you will have previous experience developing apps that are stimulating and intuitive. You will have:

• A comprehensive technical knowledge of HTML5, CSS3 (Bootstrap 3), php, Mysql, JQuery and core JavaScript languages to create an app that functions across all devices
• The ability to develop mobile apps for iOS and Android
• Excellent verbal and written English communicative skills
• A portfolio of projects accomplished successfully
• A creative flare

Benefits include cash payment plus equity share.

SUBTAP is currently receiving incubator support and is the winner of a Do It Award from UnLtd, the foundation for social entrepreneurs.

Please get in touch if you're capable and available. Prefer UK-based.

Many thanks
Harrison